refactor(filesystem): split commands.rs into per-command files (#40)
Delete src-tauri/src/filesystem/commands.rs and replace with three
single-responsibility files:
  - list.rs  (~62 lines)  — list_dir command
  - read.rs  (~40 lines)  — read_file command
  - write.rs (~131 lines) — write_file command + TEMP_COUNTER static

mod.rs re-exports list_dir, read_file, and write_file under the
same public names so src-tauri/src/lib.rs is untouched.

All three command files are thin orchestrators over scope.rs; each
file has a single responsibility and is individually testable. No
production behavior changes — all 18 tests pass unchanged.
